+++ R/get_predictions_mlogit.R
@@ -1,7 +1,7 @@
 get_predictions_mlogit <- function(model, fitfram, ...) {
   # bind IDX to new data
   dat <- insight::get_data(model)
-  fitfram <- do.call(rbind, lapply(1:length(levels(dat$idx$id2)), function(i) {
+  fitfram <- do.call(rbind, lapply(seq_along(levels(dat$idx$id2)), function(i) {
     fitfram$idx <- sprintf("%g:%s", i, levels(dat$idx$id2)[i])
     fitfram
   }))