+++ R/predictions.R
@@ -32,6 +32,8 @@ select_prediction_method <- function(model_class,
     prediction_data <- get_predictions_survival(model, data_grid, ci.lvl, type, terms, ...)
   } else if (model_class == "ols") {
     prediction_data <- get_predictions_ols(model, data_grid, ci.lvl, ...)
+  } else if (model_class == "logitr") {
+    prediction_data <- get_predictions_logitr(model, data_grid, ci.lvl, ...)
   } else if (model_class %in% c("lrm", "orm")) {
     prediction_data <- get_predictions_lrm(model, data_grid, ci.lvl, linv, ...)
   } else if (model_class == "glimML") {