+++ R/utils_model_function.R
@@ -27,6 +27,7 @@
 get_predict_function <- function(model) {
   if (inherits(model, c("wblm", "wbm"))) return("wbm")
   else if (inherits(model, "mclogit")) return("mclogit")
+  else if (inherits(model, "logitr")) return("logitr")
   else if (inherits(model, "averaging")) return("averaging")
   else if (inherits(model, "orm")) return("orm")
   else if (inherits(model, "mlogit")) return("mlogit")